+++ src/Traits/DotNotationTrait.php
@@ -20,11 +20,12 @@ trait DotNotationTrait
      * @param array $array The input array.
      * @param string $keys The dot-separated keys.
      * @param mixed|null $default returned if not null and If the key is not found in the array
+     * @param \Closure|null $callback called before return default value
      * @return mixed The value found in the array.
      *
      * @throws ArrayKeyNotFoundException If the key is not found in the array.
      */
-    private function getByDot(array $array, string $keys = '', mixed $default = null): mixed
+    private function getByDot(array $array, string $keys = '', mixed $default = null, \Closure $callback = null): mixed
     {
         if ($keys === '') {
             return $array;
@@ -38,7 +39,9 @@ trait DotNotationTrait
                 $result   = $result[$item];
                 $keysPath .= $item;
             } else {
-                if($default !== null){
+                if ($default !== null) {
+                    //before return the $default value callback called
+                    $this->callClosure($callback, $keys, $default, $item);
                     return $default;
                 }
                 throw new ArrayKeyNotFoundException($item, $keysPath);
@@ -54,10 +57,10 @@ trait DotNotationTrait
      * @param array $array The input array.
      * @param string $keys The dot-separated keys.
      * @param mixed $value The value to set.
-     *
+     * @param \Closure|null $callback called after set
      * @return array The modified array.
      */
-    private function setByDot(array $array, string $keys, mixed $value): array
+    private function setByDot(array $array, string $keys, mixed $value, \Closure $callback = null): array
     {
         $items   = $this->dotToArray($keys);
         $current = &$array;
@@ -68,6 +71,7 @@ trait DotNotationTrait
             $current = &$current[$item];
         }
         $current = $value;
+        $this->callClosure($callback, $keys, $value);
         return $array;
     }
 
@@ -89,13 +93,25 @@ trait DotNotationTrait
      * @param string|int $key The key to check for existence.
      * @return bool True if the key exists, false otherwise.
      */
-    public function isset(string|int $key):bool
+    public function isset(string|int $key): bool
     {
         try {
-            $this->getByDot($this->array,$key);
+            $this->getByDot($this->array, $key);
             return true;
-        }catch (ArrayKeyNotFoundException $e){
+        } catch (ArrayKeyNotFoundException $e) {
             return false;
+        }
+    }
+
+    /**
+     * @param \Closure|null $closure
+     * @param mixed ...$params
+     * @return void
+     */
+    private function callClosure(?\Closure $closure, mixed ...$params): void
+    {
+        if ($closure instanceof \Closure) {
+            $closure(...$params);
         }
     }
 }