+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.0.2] - 2023-10-09
+
+### Added
+
+- add `callback` for `get` method
+- add `callback` for `set` method
+
+### Changed
+
+- change `get` method params and add `callback` param
+- change `set` method param and add `callback` param
+- Modify the `getByDot` method in DotNotationTrait and add `callback` param
+- Modify the `getByDotMulti` method in MultiDotNotationTrait and add `callback` param
+
+### Fixed
+
+- No Fixed
 
 ## [1.0.1] - 2023-10-08